Editor Tool at Cursor Position
- Implemented cursor position open and close functionality for the, soon to be, editor feature search tool.

+++ Engine/Engine/Renderer/GraphicsObjects/GraphicsObjectManager.cpp
@@ -476,11 +476,11 @@ void GraphicsObjectManager::ToggleGraphicsObjectDraw(GraphicsObject* const graph
 {
 	auto findGraphicsObjectAndRemoveFromDrawVector = [](GraphicsObject* const graphicsObject, const ObjectTypes::GraphicsObjectType& type)
 	{
-		unsigned int index = 0;
+		size_t index = 0;
 		auto inCollectionLambda = [&index, graphicsObject](const std::vector<GraphicsObject*>& graphicsObjectContainer) -> bool
 		{
-			unsigned int size = graphicsObjectContainer.size();
-			for (unsigned int i = 0; i < size; i++)
+			size_t size = graphicsObjectContainer.size();
+			for (size_t i = 0; i < size; i++)
 			{
 				if (graphicsObjectContainer[i] == graphicsObject)
 				{
